+++ src/tokens/selectors.ts
@@ -125,7 +125,7 @@ export const celoAddressSelector = createSelector(coreTokensSelector, (tokens) =
 })
 
 export const swappableTokensSelector = createSelector(tokensByUsdBalanceSelector, (tokens) => {
-  return tokens.filter((tokenInfo) => tokenInfo.isSwappable === true)
+  return tokens.filter((tokenInfo) => tokenInfo.isSwappable)
 })
 
 export const tokensByCurrencySelector = createSelector(
